requirements.txt
boto3
openai
python-dotenv